requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
automated-interpretability==0.0.15
babe==0.0.7
beartype==0.14.1
better-abc==0.0.3
blake3==1.0.6
blobfile==2.1.1
boostedblob==0.15.6
cachetools==6.2.0
cbor2==5.7.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
collectibles==0.1.5
colorama==0.4.6
compressed-tensors==0.11.0
config2py==0.1.38
contourpy==1.3.2
cupy-cuda12x==13.6.0
cycler==0.12.1
datasets==2.21.0
depyf==0.19.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
dol==0.3.20
einops==0.8.1
email-validator==2.3.0
entmax==1.3
exceptiongroup==1.3.0
fancy-einsum==0.0.3
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
fastrlock==0.8.3
fasttext==0.9.3
filelock==3.19.1
flash_attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.7.4.post1/flash_attn-2.7.4.post1+cu12torch2.6cxx11abiTRUE-cp310-cp310-linux_x86_64.whl#sha256=bfdb0f290cc3d21d0810ba49a360ef91090f62cdc1345ec6900447e0d12d99af
fonttools==4.60.0
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2024.6.1
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
gprof2dot==2025.4.14
graze==0.1.31
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.35.1
i2==0.1.48
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
interegular==0.3.3
jaxtyping==0.3.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonargparse==3.13.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lark==1.2.2
lightning-utilities==0.15.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
lxml==4.9.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.8.5
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
narwhals==2.5.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.108.2
openai-harmony==0.0.4
opencv-python-headless==4.12.0.88
orjson==3.11.3
outlines_core==0.2.11
packaging==25.0
pandas==2.3.2
partial-json-parser==0.2.1.1.post6
patsy==1.0.1
pillow==11.3.0
platformdirs==4.4.0
plotly==5.24.1
plotly-express==0.4.1
pluggy==1.6.0
portalocker==3.2.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.3.2
protobuf==4.25.8
psutil==7.1.0
py-cpuinfo==9.0.0
py2store==0.1.21
pyarrow==21.0.0
pybase64==1.4.2
pybind11==3.0.1
pycountry==24.6.1
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
pytest-profiling==1.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-lightning==2.5.6
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.0.0
ray==2.49.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
sacrebleu==2.5.1
sae-bench==0.4.2
sae-lens==4.4.5
safetensors==0.4.5
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.1
sentry-sdk==2.38.0
setproctitle==1.3.7
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.48.0
statsmodels==0.14.5
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
tomli==2.2.1
torch==2.8.0
torchaudio==2.8.0
torchmetrics==0.10.3
torchvision==0.23.0
tqdm==4.67.1
traitlets==5.14.3
transformer-lens==2.16.1
transformers==4.56.2
transformers-stream-generator==0.0.5
triton==3.4.0
typeguard==4.4.4
typer==0.12.5
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
unbabel-comet==2.2.7
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
vllm==0.10.2
wadler_lindig==0.1.7
wandb==0.22.0
watchfiles==1.1.0
websockets==15.0.1
xformers==0.0.32.post1
xgrammar==0.1.23
xxhash==3.5.0
yarl==1.20.1
zstandard==0.22.0
